+++ changelog.md
@@ -1,5 +1,15 @@
 # BuzzingPixel DataModel Changelog
 
+## 1.2.0
+
+### New
+
+- Added DateTimeZone model data type
+
+### Fixed
+
+- Fixed an issue where an integer would not be accepted as a timestamp for setting DateTime data type
+
 ## 1.1.0
 
 ### New